Adjust build configuration
Adjust build configuration to make the macposts library a bit more portable and
standalone. Also simplify the CMakeLists.txt file a bit.

* CMakeLists.txt: adjust how the Snap library is used. Now it is directly
compiled into the macposts library, so that all the definitions are available in
macposts. This may not be tasteful, but could save a lot of troubles considering
that the macposts library itself is not very well structured. Also, remove
unnecessary variables to simplify this file a bit.