Remove advds
Remove the advds library. This library is rarely used only and has some license
issues. So we remove it.

* lib/advds: remove library.

* src/gridlock_checker.h: do not include the now removed "union_find.h" file. It
is also not used anywhere.

* src/shortest_path.cpp: do not include the now removed "radix_heap.h" file. It
is also unused.

* src/k_shortest_path.h:, src/k_shortest_path.cpp: remove files. This module is
not used anywhere and the implementation may be incorrect. Also it depends on
the now removed "d_array_heap.h". So remove this module as well.

* CMakeLists.txt: clean references to now removed files.