Fix build on macOS
For some reason Clang on macOS is confused by the overloaded '!=' operator used
in 'pybind11/eigen.h', which is defined in both the standard library as well as
the Snap library. As a quick and dirty first, we reorder the includes of
'pybind11/eigen.h' and 'Snap.h', so it will pick the one in the standard
library.

This probably is not a proper fix, but I do not have an environment to dig into
it. Hopefully someone else would take care of the macOS build.

* macposts/_macposts_ext.cpp: reorder includes to avoid a compilation error on
macOS. Also apply minor stylistic improvements.