Finish release workflow configuration
Currently we only build the wheels for CPython on the native architecture. Also,
because Snap does not build against musl, so we will stick with manylinux now.

Besides, we now also upload the artifacts to GitHub assets for easy download.

* .github/workflows/release.yml: fix cibuildwheel configuration; add new step
for uploading artifacts to assets; make only tagged releases trigger the
workflow.