Improve data reader
Previously the data reader was too strict in terms of the format -- for example,
many files are required to have a header line and no leading and trailing empty
lines are allowed. This commit change the readers to ignore leading and trailing
empty lines, leading and trailing white spaces of each line, and lines start
with a '#' character.

I skimmed the previous projects using macposts, and think no projects should be
broken by this change. However, This is still a breaking change. Therefore, I
would like to install this patch on main in v0.3.0.

* src/io.cpp:
* src/multiclass.cpp: make the data reader ignore leading empty lines, trailing
empty lines, and lines start with a '#' character.