Add basic testing
Currently the test cases only test if the application could run as a whole. The
correctness of the library and individual modules is not checked yet.

* setup.cfg: add new development dependency `pytest' and basic configuration for
testing. Note that the current configuration of pytest is simple enough to put
here, but it is advised to use other files when the complexity grows.

* tests/__init__.py: new file.
* tests/conftest.py: new file for test setup.
* tests/test_dta.py:
* tests/test_mcdta.py: new test cases for DTA and MCDTA models.

* tests/fixtures/__init__.py: new file.
* tests/fixtures/network_threelink.py:
* tests/fixtures/network_sevenlink.py: data for test networks.