Avoid `exit' in libmacposts
As a library, it is not supposed to terminate the whole process when errors
happen. Otherwise, for example, the whole Python REPL will die if an error
occurs in the library. So we convert all such calls to `exit' to roughly
equivalent `std::runtime_errors'.

* src/dlink.cpp:
* src/dnode.cpp:
* src/factory.cpp:
* src/io.cpp:
* src/marginal_cost.cpp:
* src/multiclass.cpp:
* src/od.cpp:
* src/path.cpp:
* src/realtime_dta.cpp:
* src/routing.cpp:
* src/so_routing.cpp:
* src/statistics.cpp:
* src/vms.cpp:
* src/workzone.cpp: convert `exit' calls to runtime errors. Also clean some
source file a bit and remove some redundant comments.