Release v0.2.1
This release mainly brings a few improvements on the configuration and data
readers. Now they are less surprising. Note that the future plan is to remove
those reader completely from libmacposts, so we probably will not improve it any
further.

Also in this release we introduce some basic test cases. They cannot guarantee
the correctness of the library, but could bring a bit of confidence while
developing.

* macposts/__init__.py: bump version 0.2.1-dev2 → 0.2.1.