Fix failing tests
This is a followup to [1: 046f43b] in which we install many changes. There is
still much work left after this great merging, but this commit at least fixes
the obvious problems in the test cases.

* macposts/_macposts_ext.cpp: Accept "Adaptive" routing type for multiclass DTA
as well. Reformat.

* macposts/dta.py: Update how we get cumulative curves for all registered links.
Currently the `get_*_link_*_cc' functions return sparse results, where only the
changes are included. So we update the implementation accordingly.

* src/dta.cpp: Remove unneeded and problematic include line. Fix the off-by-one
error in `Dta::loading'. Remove some commented lines of code.

* src/multiclass.cpp: Restore changes to the reader, which got lost during the
merge.

* tests/fixtures/network_3link.py:
* tests/fixtures/network_7link.py: Update configuration to please the current
reader.

* tests/test_dta.py:
* tests/test_mcdta.py: Update test cases to match the current interface.

1: 2023-03-23 046f43b45872e708ff86cef009e8a0ebd20c36df
   Install accumulated changes