Clean up `MNM_DTA::loading'
As a followup to the previous commit, we clean up the `MNM_DTA::loading'
function a bit. Notably, we use the C++ stream IO instead of the C style IO
function, and use more primitive data types instead of the Snap ones, as the
initial step to drop the Snap dependency.

* src/dta.cpp: Use C++ stream IO; Use primitive data types; Reformat.