Support looking up nodes/links with data
In the existing codebase, we often need to get a link/node by an ID instead of a
reference to the object. Therefore, it would be less frictional to move to this
new module if we make it easier to work with IDs.

* src/graph.cpp:
* src/graph.h: Store nodes/links with `unordered_map' using data as the key. Add
methods to look up nodes/links with data, and a method to add links with node
data. Adapt other methods/functions to this change.