Drop all `noexcept' specifiers for now
I am not sure if it is worth the effort to correctly mark all non-throwing
functions as `noexcept' (and honestly I think it is too much work when we need
to use many external things -- also compilers are not helpful for that).
However, it seems to me that a wrong `noexcept' is more harmful than a missing
`noexcept' because the former would terminate the process for exceptions. So let
us just drop all `noexcept' specifiers for now, and maybe re-add them in the
future.

* src/graph.cpp:
* src/graph.h: Drop `noexcept' specifiers.