Fix include of the graph module
I forgot how compilers deal with C++ templates. Anyway, this is quite
problematic and is just to fix the failing CI. We need to consider how should we
implement the graph class. Maybe we should just put all the definitions in the
header.

* CMakeLists.txt: Remove 'src/graph.cpp' for file list -- this file should be
included directly so that templates can be instantiated.

* src/shortest_path.cpp: Include 'graph.cpp' instead of 'graph.h'.