Refine use of `const'
Because we use `unique_ptr', we could actually modify the nodes or links even
though the graph is marked as `const'. However, that does not align with our
memory model: graph owns all nodes and links. So in this commit we change to
disallow modification through collections (container of iterators) if the graph
itself is readonly.

This adds a bunch of new overloads and slows the compilation more, but IMO we
have to live with this.

* src/graph.h: Correct use of `const' on graph methods.
* src/graph.cpp: Add new test cases for `const'.