Force a flush after network loading
We extensively use C++ IO functions (`iostream' and `cstdio') in the extension
for logging, which is frankly very distasteful to me. So I added the
`macposts.utils.silence' helper to suppress such messages.

However, because of the IO buffer, we may not be able to capture/redirect all
outputs. So we often still see some of such messages in batch mode even with
this helper (or with the pytest's output capturing). One common symptom is that
we see some random text after pytest's report.

That is quite minor an issue, but recently it has gone beyond the point that I
can tolerate while I am debugging some reproducibility issues. Therefore, I
decide to force a flush of the output buffers after network loading. This is
only a workaround, and does not cover all cases. As noted in the comments, we
should eventually fix those problems in a better way.

* macposts/_ext/dta.cpp:
* macposts/_ext/mcdta.cpp: Force a flush of buffers after loading.